requirements.txt
torch>=1.13.1
torchvision>=0.14.1
numpy
Pillow>=10.1.0
whatis
typing
diffusers>=0.30.3
av>=10.0.0
argparse>=1.4.0
packaging
future